Initial commit in branch mutiarch
In this branch squire will be rewritten to easilly support multiple
architectures. Therefore autotools is used instead of plain makefiles
and the usage of TCC is dropped. Piece by piece more of squire will be
ported to this structure and the architecture abstraction is partly
rewritten to support SMP later on

Signed-off-by: Joppe Blondel <joppe@blondel.nl>